requirements.txt
Django>=4.2,<5.0
Pillow>=10.0.0
djangorestframework>=3.16.0
django-cors-headers>=4.9.0
psycopg2-binary>=2.9.0
cloudinary>=1.44.0
python-decouple>=3.8
django-filter>=25.2
djangorestframework-simplejwt>=5.5.0
dj-database-url>=3.0.0
